Updated isLicensed() function
Original code relies on calling an advanced level tool, so code now checks if there is a valid desktop license and if it is ArcView/ArcEditor checks that spatial analyst extension is available. If these scenarios are not available then the tool is considered as unlicensed and will not run.