GeoData Toolbox containing many tools
This is the Toolbox created by Duncan. It contains many tools for computing the various metrics required by the confinement project. They are design to use the output of the confinement tool (developed by Joe's team) but to also be as generic as possible. They are built around the intention of using them on UK data and whilst I have tried to be as generic as possible some "lock in assumptions" may have crept in. Don't know until tested!