New tools added and general changes to toolbox
As well as adding new tools, I have ensured all tools are pointing to the help file which must be stored here: C:\GISApplications\Confinement Tools Help.chm

Tools are now numbered so it is more obvious in which order you should complete them. But some tools rely on external tools being run first and these are indicated in the workflow diagram in the associated help file.